Resolved router utils dependency cycle
