+++ client/app/scripts/components/app.js
@@ -42,7 +42,8 @@ import TimeTravelWrapper from './time-travel-wrapper';
 import ViewModeSelector from './view-mode-selector';
 import NetworkSelector from './networks-selector';
 import DebugToolbar, { showingDebugToolbar, toggleDebugToolbar } from './debug-toolbar';
-import { getRouter, getUrlState } from '../utils/router-utils';
+import { getUrlState } from '../utils/router-utils';
+import { getRouter } from '../router';
 import { trackAnalyticsEvent } from '../utils/tracking-utils';
 import { availableNetworksSelector } from '../selectors/node-networks';
 import { timeTravelSupportedSelector } from '../selectors/time-travel';
